Refactor monolithic services.py into separate modules per PCO product
- server.py: Auth, MCP app setup, shared helpers (_build_patch_body), entrypoint
- services.py: All Services API tools (service types, plans, items, team members, schedules, songs, tags)
- registrations.py: All Registrations API tools (events, categories, event times, attendees, duplication)
- Updated Dockerfile to copy all modules and use server.py as entrypoint
